changes how plugins are loaded

+++ python/index.py
@@ -47,13 +47,17 @@ def import_lib(url):
     else: print("- the option is always available -")
 
 def import_plugin(url):
-    local_input = input("- this will download the python file for the plugin and run it, please be safe with the code you download, would you like to continue? (y/n) -")
-    if local_input.lower() == "y":
-        file_name = "plug.py"
-        urllib.request.urlretrieve(url, file_name)
-        globals()["plug"] = __import__("plug")
+    global plugin_is_active
+    if (plugin_is_active is False):
+        local_input = input("- this will download the python file for the plugin and run it, please be safe with the code you download, would you like to continue? (y/n) -")
+        if local_input.lower() == "y":
+            file_name = "plug.py"
+            urllib.request.urlretrieve(url, file_name)
+            globals()["plug"] = __import__("plug")
 
-    else: print("- the option is always available -")
+        else: print("- the option is always available -")
+    else:
+        print("sorry but a plugin is already loaded, please exit the script to load a different one")
 
 #commands
 def check(ci):
@@ -84,12 +88,15 @@ def check(ci):
     #allows plugins to be inputted
     elif ci == ".plugin":
         global plugin_is_active
-        local_input = input("please input the url to your plugin - ")
-        import_plugin(local_input)
-        print("loading plugin data...")
-        command_Text.update(plug.command_Text)
-        plugin_is_active = True
-        print("plugin loaded, to check new content run '?help' to list all commands")
+        if (plugin_is_active is False):
+            local_input = input("please input the url to your plugin - ")
+            import_plugin(local_input)
+            print("loading plugin data...")
+            command_Text.update(plug.command_Text)
+            plugin_is_active = True
+            print("plugin loaded, to check new content run '?help' to list all commands")
+        else:
+            print("sorry but a plugin is already loaded, please exit the script to load a different one")
 
     elif ci == ".local_plugin":
         local_input = input("please enter the name and location of the file -")
